applications:aoa_locator:Fixes wrong number of sampling slots provided in df_lib config
Fixes number of sampling slots provided to df_library in configuration entry sampling_slots_num.
The value is wrong when oversampling is enabled due to fact that samples collected
in switch slots are removed before angles evaluation. In that case number of slots
is half of effective number of slots.

Jira:DEB-171

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>